refactor(SidebarSearch): Move filters to SidebarSearch comp
